Rename 'testmocking' to 'loadtesting'
