+++ loadtesting/management/commands/clear_mock_data.py
@@ -10,7 +10,7 @@ from django.conf import settings
 from guidedmodules.models import AppSource, Module
 from siteapp.models import User, Organization
 
-from testmocking.data_management import delete_objects
+from loadtesting.data_management import delete_objects
 
 class Command(BaseCommand):
     help = 'Clear all data created for mock purposes'
